added folders to split up backend calls. Auth folder is started. Plan to rewrite and check functions as we organize them
