+++ frontend/src/api/backend_calls.js
@@ -8,45 +8,6 @@ async function basicFetch(url, payload) {
     return body
   }
 
-
-  export async function signup(context) {
-  
-    const payload = {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(context)
-    }
-    const response = await basicFetch(`http://${host}/api/v1/accounts/register/signup`,payload)
-    console.log(response.status, "body")
-    return response
-  }
-
-  export async function login(context) {
-    const payload = {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(context)
-    }
-    const body = await basicFetch(`http://${host}/api/v1/accounts/register/get-token`, payload)
-    return body.token
-  }
-
-  export async function emailResetLink(context) {
-    const payload = {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(context)
-    }
-    const response = await basicFetch(`http://${host}/api/v1/password_reset/`, payload)
-    return response
-  }
-
   export async function emailResetConfirm(context) {
     const payload = {
       method: "POST",
@@ -192,7 +153,6 @@ export async function joinTeam(context){
       headers: {
         "Content-Type": "application/json",
         "Authorization": `Token ${localStorage.getItem("token")}`
-
       }
     })
     return response
