+++ frontend/src/components/DashboardLeagues.jsx
@@ -1,7 +1,7 @@
 import { useContext } from "react"
 import { useNavigate } from "react-router-dom";
 import UserContext from "../contexts/UserContext"
-import { useNavigate} from 'react-router-dom'
+
 
 
 export default function DashboardLeagues(){
@@ -13,11 +13,6 @@ export default function DashboardLeagues(){
         navigate(`/league/${league_id}`);
     };
 
-    const navigate = useNavigate();
-    
-    const handleViewLeague = (league_id) => {
-        navigate(`/league/${league_id}`);
-    };
 
     console.log(usersTeams)
